Switch to hosts next, remove old consul code
